restructured Person as trait
